Revert "graphs fixed"
This reverts commit a502c41c60d861a3420a4670d24a3a4362a0dc2e.

+++ DTM-WPF/DTM-WPF/GraphBits/PocVertex.cs
@@ -13,14 +13,14 @@ namespace DTM_WPF{
     {
         public int serviceId { get; set; }
         public string serviceName { get; set; }
-        public int perc { get; set; }
+        public float perc { get; set; }
         public string color { get; set; }
 
         public PocVertex(string servicename, int serviceid, float p, string Color)
         {
             serviceName = servicename;
             serviceId = serviceid;
-            perc = (int)p;
+            perc = p;
             color = Color;
         }
